requirements.txt
langchain
langchain-openai
chromadb
PyMuPDF
tqdm
streamlit
python-dotenv
